[package] Make it possible to re-save a PackageImporter module
As title. Previously this was guarded against for implementation
simplicity, as we didn't really think there was a use case for saving a
mangled module name directly.

But people started doing stuff like:
```
exporter.save_module(my_imported_obj.__module__)
```
which implicitly passes along the mangled module name.

This PR makes it so that given `PackageImporter` instance can always
import modules that it created, and changes `PackageExporter` to
properly demangle the resulting module name when writing the package to
the export archive.

ghstack-source-id: 1ff2244085f44476dd725589e8500cf530d8d330
Pull Request resolved: https://github.com/pytorch/pytorch/pull/65101